Add support for targeting an existing page list
If a designer has already created a page list and wants to use this as
the page list with SnapSlider, it is possible to do this by specifying
a selector for the elements which should change pages. This is usually
a collection of anchors, but it can be any element that supports having
a click event attached. The number of links must match the number of
items in the SnapSlider, or the initialisation will halt and a warning
logged to the console.