Enable three-stage image replacement, dependent on state of slider
By default, you should load your backgrounds with the lowest file-size
versions of your images. These are known as the 'normal' state. There
are then two additional states - 'large' and 'largest' - which will be
replaced under the following circumstances:

- When the slider is not zoomed in, large images will be loaded for the
  elements in view, with all other elements remaining normal.
- When the slider is zoomed in, largest images will be loaded for the
  elements in view, with all other elements getting large images.

When this is merged into master, I'll update the gh-pages branch with
an example of this feature being used.